test type conversion

+++ Avatar_Micro_Robot/Robot_V2/Power_Board/trunk/closed_loop_control/core/InputCapture.c
@@ -289,7 +289,7 @@ void IC_Init(const kICModule module,
 }
 
 
-float IC_period(const kICModule module) {
+unsigned int IC_period(const kICModule module) {
   return periods[module];
 }
